Refactor: Move log types to a different file and update imports
