+++ app/client/src/entities/AppsmithConsole/logtype.ts
@@ -1,0 +1,9 @@
+enum LOG_TYPE {
+  WIDGET_PROPERTY_VALIDATION_ERROR,
+  WIDGET_UPDATE,
+  ACTION_EXECUTION_ERROR,
+  ACTION_EXECUTION_SUCCESS,
+  ENTITY_DELETED,
+}
+
+export default LOG_TYPE;
