+++ app/client/src/sagas/ActionExecutionSagas.ts
@@ -105,7 +105,8 @@ import {
   resetWidgetMetaProperty,
 } from "actions/metaActions";
 import AppsmithConsole from "utils/AppsmithConsole";
-import { ENTITY_TYPE, LOG_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import LOG_TYPE from "entities/AppsmithConsole/logtype";
 
 export enum NavigationTargetType {
   SAME_WINDOW = "SAME_WINDOW",
