+++ app/client/src/sagas/WidgetOperationSagas.tsx
@@ -112,7 +112,8 @@ import {
   WIDGET_DELETE,
 } from "constants/messages";
 import AppsmithConsole from "utils/AppsmithConsole";
-import { ENTITY_TYPE, LOG_TYPE } from "entities/AppsmithConsole";
+import { ENTITY_TYPE } from "entities/AppsmithConsole";
+import LOG_TYPE from "entities/AppsmithConsole/logtype";
 
 function* getChildWidgetProps(
   parent: FlattenedWidgetProps,
